Status code for trying to lock obj locked by other user should be 409
Previously, this status code was listed as 401 which is not correct.

Status Code 401:
Similar to 403 Forbidden, but specifically for use when authentication
is possible but has failed or not yet been provided. The response must
include a WWW-Authenticate header field containing a challenge
applicable to the requested resource

Status Code 409:
Indicates that the request could not be processed because of conflict
in the request, such as an edit conflict.