Improved API: fixed edgewise deletion bug, better status codes, etc
Now delete only works on locks controlled by that user. This prevents
and edge-case bug where a user loses a lock and then refreshes their
browser before the ping that would notify the JS of this change. The
unload would send a delete signal and then they would regain the lock
when the page reloaded.

If a user needs to delete a lock that they don’t control, they can take
control of that lock with a PUT request before DELETING it.

Additionally, response codes that return no data are now 204 and 200
response codes have a JSON response.