Fix issue with unlocking forms on page unload
Modern browsers no longer allow AJAX requests to hold up the page unload event. To enable automatically unlocking the form when the user navigates away from it, we now need to use the navigator.sendBeacon method. Unfortunately, this method only allows POST data so it can't make use of our existing RESTful API methods (which would expect a DELETE to remove a lock). As a result, this commit creates a new endpoint for delete beacon calls that mimics the behavior of the existing DELETE API REST call.